requirements.txt
alembic==1.14.0
cffi==1.17.1
cfgv==3.4.0
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
greenlet==3.1.1
identify==2.6.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightgbm==4.5.0
Mako==1.3.7
MarkupSafe==2.1.5
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
opencv-python==4.10.0.84
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
pre_commit==4.0.1
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sounddevice==0.5.1
soundfile==0.12.1
SQLAlchemy==2.0.36
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.1
